Added Curl Call to digits.
It barely works. But... The system was working as a whole physically
sorting coins for the first time.

It looks like the fileWatcher thread is dying and there is timing issues
now that the digits look up is in place. This needs to be one program,
all in C#, instead of two, and not saving the images to disk.